delete alumni implemented

+++ next/app/about/alumni/DeleteAlumni.tsx
@@ -4,19 +4,18 @@ import { useEffect, useState } from "react";
 import { AlumniMember } from "./alumni";
 
 interface DeleteAlumniProps {
-    open: boolean;
-    alumniMember?: AlumniMember;
-    fetchData: () => Promise<void>;
+    open: boolean; // open - State of delete form modal
+    alumniMember?: AlumniMember; // alumniMember - Currently selected alumni to be edited
+    fetchData: () => Promise<void>; // getAlumni - Function to get active alumni, used to update the list
     closeModal: () => void // closeModal - Function to close the form's modal
 }
 
 export const DeleteAlumniButton: React.FC<DeleteAlumniProps> = ({ open, alumniMember, fetchData, closeModal  }) => {
-    const [alumni_id, setAlumniID] = useState("");
+    // const [alumni_id, setAlumniID] = useState("");
 
     const [formData, setFormData] = useState({
         alumni_id: ''
 	});
-
 	const [error, setError] = useState("")
 
     // Fill form with current 
@@ -52,17 +51,25 @@ export const DeleteAlumniButton: React.FC<DeleteAlumniProps> = ({ open, alumniMe
 	};
 
 	const handleDelete = async () => {
+        setError("");
 		try {
-		const response = await fetch("/api/alumni", {
-			method: "DELETE",
-			body: JSON.stringify({ id: alumni_id }),
-		});
 
-		if (response.ok) {
-			handleCancel();
-			(document.getElementById("delete-alumni") as HTMLDialogElement).close();
-			fetchData();
-		}
+            if (!alumniMember?.alumni_id) {
+                    setError("Misisng alumni_id for this alumni.");
+                    return;
+            }
+
+            // Call to alumni route to delete alumni's user data
+            const response = await fetch("/api/alumni", {
+                method: "DELETE",
+                body: JSON.stringify({ id: alumniMember.alumni_id }),
+            });
+
+            if (response.ok) {
+                handleCancel();
+                (document.getElementById("delete-alumni") as HTMLDialogElement).close();
+                fetchData();
+            }
 		} catch (error) {}
 	};
